Extract NFTs grid flash list props to hook
